// frontend/src/domain/barcode/barcode.generated.ts
/* eslint-disable */
// THIS FILE IS GENERATED by codegen/plugin.js — do not edit by hand.
import type { TypedDocument } from "../../api/graphql";

export type BarcodeFragment = {
  id: string;
  gtin: string;
  itemId: string;
  packSize: number | null;
};

export type BarcodeByGtinVariables = {
  storeId: string;
  gtin: string;
};

export type BarcodeByGtinResult = {
  barcodeByGtin: ({
  __typename: "BarcodeNode";
} & BarcodeFragment) | ({
  __typename: "NodeError";
} & {
  error: {
  __typename: "DatabaseError" | "RecordNotFound";
  description: string;
};
});
};

export const BarcodeByGtin = {
  query: "query barcodeByGtin($storeId: String!, $gtin: String!) {\n  barcodeByGtin(storeId: $storeId, gtin: $gtin) {\n    __typename\n    ... on BarcodeNode {\n      ...Barcode\n    }\n    ... on NodeError {\n      error {\n        __typename\n        description\n      }\n    }\n  }\n}\n\nfragment Barcode on BarcodeNode {\n  id\n  gtin\n  itemId\n  packSize\n}",
} as TypedDocument<BarcodeByGtinResult, BarcodeByGtinVariables>;

export type InsertBarcodeVariables = {
  storeId: string;
  input: {
    gtin: string;
    itemId: string;
    packSize?: number | null;
  };
};

export type InsertBarcodeResult = {
  insertBarcode: ({
  __typename: "BarcodeNode";
} & BarcodeFragment);
};

export const InsertBarcode = {
  query: "mutation insertBarcode($storeId: String!, $input: InsertBarcodeInput!) {\n  insertBarcode(storeId: $storeId, input: $input) {\n    __typename\n    ... on BarcodeNode {\n      ...Barcode\n    }\n  }\n}\n\nfragment Barcode on BarcodeNode {\n  id\n  gtin\n  itemId\n  packSize\n}",
} as TypedDocument<InsertBarcodeResult, InsertBarcodeVariables>;